Day 14/modular-pinecone-rag-backend/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
openai
pinecone
pypdf
